Moving all of the tests over.
